% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kordle.R
\docType{package}
\name{kordle}
\alias{kordle}
\title{This package is used to play Wordle in Korean}
\description{
The package kordle has three functions.
}
\section{Check Word}{

\emph{check_word} compares the guess word to target word and provides
feedback on each character
}

\section{Output Word}{

\emph{output_word} a follow up function to use the feedback on each
character and provide feedback on the guess word
}

